agentic/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
kiteconnect==4.1.0
websockets==11.0.3
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 